Web UI katmanında Banner,Contact,Choose,Appointment Dtoları servisleri ve Controllerları yapıldı

+++ Arkitektur.WebUI/Services/ProjectServices/ProjectService.cs
@@ -14,10 +14,18 @@ namespace Arkitektur.WebUI.Services.ProjectServices
             {
                 throw new ApiValidationException(imageResponse.Errors);
             }
+
             projectDto.ImageUrl = imageResponse.Data.ImageUrl;
             var response = await _client.PostAsJsonAsync("projects", projectDto);
             var result = await response.Content.ReadFromJsonAsync<BaseResult<object>>();
-            return result.IsFailure ? throw new ApiValidationException(result.Errors) : result;
+            if (result.IsFailure)
+            {
+                await _fileService.DeleteFileAsync(imageResponse.Data.ImageUrl);
+                throw new ApiValidationException(result.Errors);
+            }
+
+
+            return result;
         }
 
         public async Task<BaseResult<object>> DeleteAsync(int id)
@@ -47,13 +55,22 @@ namespace Arkitektur.WebUI.Services.ProjectServices
         {
             if (projectDto.File is not null)
             {
-                await  _fileService.DeleteFileAsync(projectDto.ImageUrl);
                 var imageResponse = await _fileService.UploadFileAsync(projectDto.File);
-                projectDto.ImageUrl = imageResponse.Data.ImageUrl; 
+                if (imageResponse.IsFailure)
+                {
+                    throw new ApiValidationException(imageResponse.Errors);
+                }
+                await _fileService.DeleteFileAsync(projectDto.ImageUrl);
+                projectDto.ImageUrl = imageResponse.Data.ImageUrl;
             }
             var response = await _client.PutAsJsonAsync("projects", projectDto);
             var result = await response.Content.ReadFromJsonAsync<BaseResult<object>>();
-            return result.IsFailure ? throw new ApiValidationException(result.Errors) : result;
+            if (result.IsFailure)
+            {
+                throw new ApiValidationException(result.Errors);
+            }
+            return result;
+
         }
     }
 }
